add CartItem component
